requirements.txt
discord.py==2.3.2
waitress==2.1.2
Flask==2.2.3
gunicorn
python-dotenv
aiohttp
